# docs/v1/openapi.yml
openapi: 3.1.0
info:
  title: Ocha API
  version: 1.0.0
  description: |
    API REST pour la gestion de matcha shops Ocha, incluant la gestion des utilisateurs, produits, magasins et commandes.
    
    ## Fonctionnalités
    - **Gestion de l'authentification** : Inscription et connexion des utilisateurs avec JWT.
    - **Gestion des produits** : Création, mise à jour, suppression et récupération des produits.
    - **Gestion des magasins** : Création, mise à jour, suppression et récupération des magasins.
    - **Gestion des commandes** : Création de commandes par les utilisateurs, mise à jour du statut
    - **Statistiques des commandes** : Récupération des statistiques des commandes par utilisateur.

    ## Authentification
    - **Inscription** : les utilisateurs peuvent s'inscrire en fournissant une adresse e-mail, un mot de passe et un nom d'affichage, eventuellement un numéro de téléphone.
    - **Connexion** : les utilisateurs peuvent se connecter avec leur adresse e-mail et mot de passe pour obtenir un token JWT.
    - **Sécurité** : les endpoints protégés nécessitent un token JWT dans l'en-tête `Authorization Bearer <token>`.
    - **Rôles** : les utilisateurs ont des rôles (`user`, `admin`) qui contrôlent l'accès aux fonctionnalités.
    - **Hashage des mots de passe** : les mots de passe sont stockés de manière sécurisée en utilisant bcrypt.

    ## Code de status de réponse HTTP pouvant être rencontrés
    - `200 OK` : Requête réussie (ressource retournée).
...

...

...

...

...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Store:
      type: object
      required:
        - id
        - name
        - slug
        - email
        - address
        - location
        - is_active
        - opening_hours
        - created_at
      properties:
        id:
          type: string
          format: mongo-id
          description: Identifiant unique.
          example: "64c9f2e4f6a7b8c90d1e2f3"
        name:
          type: string
          minLength: 3
          maxLength: 50
          example: "Ocha Neuchâtel Gare"
        slug:
          type: string
          description: Identifiant lisible dans l'URL, unique.
          example: "ocha-neuchatel-gare"
        phone:
          type: string
          nullable: true
        email:
          type: string
          format: email
        address:
          type: object
          required:
            - line1
            - city
            - zipcode
            - country
          properties:
            line1:
              type: string
            city:
              type: string
            zipcode:
              type: string
            country:
              type: string
          example: {
            line1: "Avenue de la Gare 1",
            city: "Neuchâtel",
            zipcode: "2000",
            country: "Suisse"
          }
        location:
          type: object
          required:
            - type
            - coordinates
          properties:
            type:
              type: string
              enum: [Point]
            coordinates:
              type: array
              minItems: 2
              maxItems: 2
              items:
                type: number
          example: {
            type: "Point",
            coordinates: [6.9896, 46.9910]
          }
        is_active:
          type: boolean
          default: true
        opening_hours:
          type: array
          minItems: 7
          maxItems: 7
          items:
            type: array
            nullable: true
            items:
              type: string
              pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    ####################
    # ORDER
    ####################
    Order:
      type: object
      required:
        - id
        - user_id
        - store_id
        - status
        - pickup
        - total_price_chf
        - created_at
      properties:
        id:
          type: string
          format: mongo-id
          description: Identifiant unique.
          example: "64d0f3f5a7b8c9d01e2f3g4"
        user_id:
          type: string
          format: mongo-id
          description: Identifiant unique de l'utilisateur.
          example: "64a7f0c2e4b0f5d6c8a1b2c3"
        store_id:
          type: string
          format: mongo-id
          description: Identifiant unique du magasin.
          example: "64c9f2e4f6a7b8c90d1e2f3"
        status:
          type: string
          enum:
            - en préparation
            - prête
            - récupérée
          default: en préparation
        pickup:
          type: string
          format: date-time
        total_price_chf:
          type: number
          minimum: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    ####################
    # ORDER ITEM
    ####################
...
paths:
...
  /users/orders:
    get:
      tags:
        - Users
      summary: Reccupère les commandes de l'utilisateur
      description: Récupère les commandes passées par l'utilisateur actuellement authentifié.
      operationId: getMyOrders
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: store_id
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: User orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Your orders history"
                  page:
                    type: integer
                    example: 1
                  totalPages:
                    type: integer
                    example: 2
                  totalOrders:
                    type: integer
                    example: 15
                  orders:
                    type: array
                    items:
                      $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
  # ---------------------------------------
  # Products
  # ---------------------------------------
...
  /orders:
    post:
      tags:
        - Orders
      summary: Crée une nouvelle commande
      description: Permet à un utilisateur authentifié de créer une commande dans un magasin spécifique avec une date de pickup et des items.
      operationId: createOrder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - store_id
                - pickup
                - items
              properties:
                store_id:
                  type: string
                  description: ID du magasin
                pickup:
                  type: string
                  format: date-time
                  description: Date et heure de retrait
                items:
                  type: array
                  items:
                    type: object
                    required:
                      - product_id
                      - size
                      - quantity
                    properties:
                      product_id:
                        type: string
                        description: ID du produit
                      size:
                        type: string
                        enum: [S, M, L]
                      quantity:
                        type: integer
                        minimum: 1
                        default: 1
            example:
              store_id: "6962e0fd6aaa44672169d804"
              pickup: "2026-01-12T14:00:00.000Z"
              items:
                - product_id: "695fb951bf4a414b74147b9e"
                  size: "M"
                  quantity: 1
      responses:
        "201":
          description: Commande créée avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Order successfully created"
                  order:
                    type: object
                    description: Détails de la commande créée
                    properties:
                      _id:
                        type: string
                        format: mongo-id
                      user_id:
                        type: object
                        properties:
                          _id:
                            type: string
                            format: mongo-id
                          email:
                            type: string
                          display_name:
                            type: string
                      store_id:
                        $ref: "#/components/schemas/Store"
                      status:
                        type: string
                        example: "en préparation"
                      pickup:
                        type: string
                        format: date-time
                      total_price_chf:
                        type: number
                        example: 7.9
                      createdAt:
                        type: string
                        format: date-time
                      updatedAt:
                        type: string
                        format: date-time
                      __v:
                        type: integer
              example:
                message: "Order successfully created"
                order:
                  _id: "6962e54c6aaa44672169d83c"
                  user_id:
                    _id: "6960b6eb75f97d829b1978fc"
                    email: "user@ocha.ch"
                    display_name: "User_2002"
                  store_id:
                    _id: "6962e0fd6aaa44672169d804"
                    name: "Ocha Neuchâtel Gare"
                    slug: "ocha-neuchatel-gare"
                    phone: "12345678"
                    email: "neuchatel@ocha.ch"
                    is_active: true
                    address:
                      line1: "Avenue de la Gare 2"
                      city: "Neuchâtel"
                      zipcode: "2000"
                      country: "Suisse"
                    location:
                      type: "Point"
                      coordinates: [6.931444286600357, 46.99321666053446]
                    opening_hours:
                      - []
                      - ["08:00","18:00"]
                      - ["08:00","18:00"]
                      - ["08:00","18:00"]
                      - ["08:00","18:00"]
                      - ["08:00","20:00"]
                      - ["08:00","20:00"]
                    created_at: "2026-01-10T23:30:05.756Z"
                    updated_at: "2026-01-10T23:32:59.994Z"
                    __v: 1
                  status: "en préparation"
                  pickup: "2026-01-12T14:00:00.000Z"
                  total_price_chf: 7.9
                  createdAt: "2026-01-10T23:48:28.906Z"
                  updatedAt: "2026-01-10T23:48:28.917Z"
                  __v: 0
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Store not found
        "422":
          description: Validation error
        "500":
          description: Internal server error
  /orders/{id}:
    get:
      tags:
        - Orders
      summary: Récupère une commande par ID
      description: Permet à un utilisateur authentifié de récupérer une commande spécifique par son identifiant.
      operationId: getOrderById
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            description: ID de la commande
      responses:
        "200":
          description: Commande trouvée
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Order found"
                  order:
                    type: object
                    description: Détails de la commande
                    properties:
                      _id:
                        type: string
                        format: mongo-id
                      user_id:
                        type: object
                        properties:
                          _id:
                            type: string
                            format: mongo-id
                          email:
                            type: string
                          display_name:
                            type: string
                      store_id:
                        $ref: "#/components/schemas/Store"
                      status:
                        type: string
                        example: "en préparation"
                      pickup:
                        type: string
                        format: date-time
                      total_price_chf:
                        type: number
                        example: 7.9
                      createdAt:
                        type: string
                        format: date-time
                      updatedAt:
                        type: string
                        format: date-time
                      __v:
                        type: integer
              example:
                message: "Order found"
                order:
                  _id: "6962e54c6aaa44672169d83c"
                  user_id:
                    _id: "6960b6eb75f97d829b1978fc"
                    email: "user@ocha.ch"
                    display_name: "User_2002"
                  store_id:
                    _id: "6962e0fd6aaa44672169d804"
                    name: "Ocha Neuchâtel Gare"
                    slug: "ocha-neuchatel-gare"
                    phone: "12345678"
                    email: "neuchatel@ocha.ch"
                    is_active: true
                    address:
                      line1: "Avenue de la Gare 2"
                      city: "Neuchâtel"
                      zipcode: "2000"
                      country: "Suisse"
                    location:
                      type: "Point"
                      coordinates: [6.931444286600357, 46.99321666053446]
                    opening_hours:
                      - []
                      - ["08:00","18:00"]
                      - ["08:00","18:00"]
                      - ["08:00","18:00"]
                      - ["08:00","18:00"]
                      - ["08:00","20:00"]
                      - ["08:00","20:00"]
                    created_at: "2026-01-10T23:30:05.756Z"
                    updated_at: "2026-01-10T23:32:59.994Z"
                    __v: 1
                  status: "en préparation"
                  pickup: "2026-01-12T14:00:00.000Z"
                  total_price_chf: 7.9
                  createdAt: "2026-01-10T23:48:28.906Z"
                  updatedAt: "2026-01-10T23:48:28.917Z"
                  __v: 0
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Order not found
        "500":
          description: Internal server error

    patch:
      tags:
        - Orders
      summary: Mettre à jour le statut d'une commande
      description: |
        Permet de mettre à jour le statut d'une commande existante.
        Route généralement réservée aux administrateurs ou au personnel du magasin.
      operationId: updateOrderStatus
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          description: Identifiant de la commande
          schema:
            type: string
            format: mongo-id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  description: Nouveau statut de la commande
                  example: "prête"
            example:
              status: "prête"
      responses:
        "200":
          description: Statut de la commande mis à jour
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Order status updated"
                  order:
                    type: object
                    properties:
                      _id:
                        type: string
                        format: mongo-id
                      user_id:
                        type: string
                        format: mongo-id
                      store_id:
                        type: string
                        format: mongo-id
                      status:
                        type: string
                        example: "prête"
                      pickup:
                        type: string
                        format: date-time
                      total_price_chf:
                        type: number
                        example: 7.9
                      createdAt:
                        type: string
                        format: date-time
                      updatedAt:
                        type: string
                        format: date-time
                      __v:
                        type: integer
              example:
                message: "Order status updated"
                order:
                  _id: "6962e54c6aaa44672169d83c"
                  user_id: "6960b6eb75f97d829b1978fc"
                  store_id: "6962e0fd6aaa44672169d804"
                  status: "prête"
                  pickup: "2026-01-12T14:00:00.000Z"
                  total_price_chf: 7.9
                  createdAt: "2026-01-10T23:48:28.906Z"
                  updatedAt: "2026-01-10T23:55:04.507Z"
                  __v: 0
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Order not found
        "500":
          description: Internal server error

    delete:
      tags:
        - Orders
      summary: Supprime une commande
      description: Supprime une commande existante par son identifiant.
      operationId: deleteOrder
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Order deleted successfully
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Order not found
        "500":
          description: Internal server error
...